[Fleet] Improve fleet package policy delete API (#140541)

+++ x-pack/plugins/fleet/server/services/package_policy.ts
@@ -9,7 +9,6 @@ import { omit, partition, isEqual } from 'lodash';
 import { i18n } from '@kbn/i18n';
 import semverLt from 'semver/functions/lt';
 import { getFlattenedObject } from '@kbn/std';
-import { SavedObjectsErrorHelpers } from '@kbn/core/server';
 import type {
   KibanaRequest,
   ElasticsearchClient,
@@ -23,6 +22,8 @@ import { safeLoad } from 'js-yaml';
 import { DEFAULT_SPACE_ID } from '@kbn/spaces-plugin/common/constants';
 
 import type { AuthenticatedUser } from '@kbn/security-plugin/server';
+
+import pMap from 'p-map';
 
 import {
   packageToPackagePolicy,
@@ -552,42 +553,54 @@ class PackagePolicyService implements PackagePolicyServiceInterface {
   ): Promise<DeletePackagePoliciesResponse> {
     const result: DeletePackagePoliciesResponse = [];
 
-    for (const id of ids) {
+    const packagePolicies = await this.getByIDs(soClient, ids, { ignoreMissing: true });
+    if (!packagePolicies) {
+      return [];
+    }
+
+    const uniqueAgentPolicyIds = [
+      ...new Set(packagePolicies.map((packagePolicy) => packagePolicy.policy_id)),
+    ];
+
+    const hostedAgentPolicies: string[] = [];
+
+    for (const agentPolicyId of uniqueAgentPolicyIds) {
       try {
-        const packagePolicy = await this.get(soClient, id);
+        await validateIsNotHostedPolicy(
+          soClient,
+          agentPolicyId,
+          options?.force,
+          'Cannot remove integrations of hosted agent policy'
+        );
+      } catch (e) {
+        hostedAgentPolicies.push(agentPolicyId);
+      }
+    }
+
+    const deletePackagePolicy = async (id: string) => {
+      try {
+        const packagePolicy = packagePolicies.find((p) => p.id === id);
+
         if (!packagePolicy) {
-          throw new Error('Package policy not found');
+          throw new PackagePolicyNotFoundError(
+            `Saved object [ingest-package-policies/${id}] not found`
+          );
         }
 
         if (packagePolicy.is_managed && !options?.force) {
           throw new PackagePolicyRestrictionRelatedError(`Cannot delete package policy ${id}`);
         }
 
-        await validateIsNotHostedPolicy(
-          soClient,
-          packagePolicy?.policy_id,
-          options?.force,
-          'Cannot remove integrations of hosted agent policy'
-        );
-
-        const agentPolicy = await agentPolicyService
-          .get(soClient, packagePolicy.policy_id)
-          .catch((err) => {
-            if (SavedObjectsErrorHelpers.isNotFoundError(err)) {
-              appContextService
-                .getLogger()
-                .warn(`Agent policy ${packagePolicy.policy_id} not found`);
-              return null;
-            }
-            throw err;
-          });
-
-        await soClient.delete(SAVED_OBJECT_TYPE, id);
-        if (agentPolicy && !options?.skipUnassignFromAgentPolicies) {
-          await agentPolicyService.bumpRevision(soClient, esClient, packagePolicy.policy_id, {
-            user: options?.user,
-          });
+        if (hostedAgentPolicies.includes(packagePolicy.policy_id)) {
+          throw new HostedAgentPolicyRestrictionRelatedError(
+            'Cannot remove integrations of hosted agent policy'
+          );
         }
+
+        // TODO: replace this with savedObject BulkDelete when following PR is merged
+        // https://github.com/elastic/kibana/pull/139680
+        await soClient.delete(SAVED_OBJECT_TYPE, id);
+
         result.push({
           id,
           name: packagePolicy.name,
@@ -605,6 +618,25 @@ class PackagePolicyService implements PackagePolicyServiceInterface {
           success: false,
           ...ingestErrorToResponseOptions(error),
         });
+      }
+    };
+
+    await pMap(ids, deletePackagePolicy, { concurrency: 1000 });
+
+    if (!options?.skipUnassignFromAgentPolicies) {
+      const uniquePolicyIdsR = [
+        ...new Set(result.filter((r) => r.success && r.policy_id).map((r) => r.policy_id!)),
+      ];
+
+      const agentPolicies = await agentPolicyService.getByIDs(soClient, uniquePolicyIdsR);
+
+      for (const policyId of uniquePolicyIdsR) {
+        const agentPolicy = agentPolicies.find((p) => p.id === policyId);
+        if (agentPolicy) {
+          await agentPolicyService.bumpRevision(soClient, esClient, policyId, {
+            user: options?.user,
+          });
+        }
       }
     }
